move filter functions to new folder structure
