+++ matchms/filtering/metadata_processing/derive_formula_from_name.py
