+++ matchms/filtering/metadata_processing/repair_smiles_from_compound_name.py
