+++ matchms/filtering/peak_processing/reduce_to_number_of_peaks.py
