+++ matchms/filtering/peak_processing/require_minimum_of_high_peaks.py
